server/requirements.txt
FLask
flask-cors
Flask-WTF
python-dotenv
